Added Sample code for Ingestion API
